requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.5
annotated-types==0.7.0
banal==1.0.6
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
dataset==1.6.2
easyocr==1.7.0
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
huggingface-hub==0.34.4
idna==3.4
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.7
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
Levenshtein==0.27.1
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.9.4
modules==1.0.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
ninja==1.13.0
nltk==3.9.1
numpy==1.26.2
nvidia-ml-py==12.575.51
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
optuna==4.5.0
packaging==25.0
pandas==2.2.2
Pillow==10.1.0
platformdirs==4.4.0
protobuf==6.32.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pynvml==12.0.0
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.1
RapidFuzz==3.13.0
regex==2025.9.1
requests==2.28.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.13.1
sentry-sdk==2.35.2
shapely==2.0.7
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
SQLAlchemy==1.4.54
stringzilla==4.0.7
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.19
tomli==2.2.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
utils==1.0.2
wandb==0.21.3
zipp==3.23.0